move the implementation to its own class as per review
